// src/ts/Enka.Network_icon2text/types/languages.ts
export type languages
    = ""
    | "DE" // German
    | "EN" // English
    | "ES" // Spanish
    | "FR" // French
    | "ID" // Indonesian
    | "IT" // Italian
    | "PT" // Portuguese
    | "RU" // Russian
    | "TH" // Thai
    | "TR" // Turkish
    | "VI" // Vietnamese
    | "JA" // Japanese
    | "KO" // Korean
    | "ZH-CH" // Simplified Chinese
    | "ZH-TW"; // Traditional Chinese